+++ src/layout/houses/shape.rs
@@ -54,9 +54,9 @@ const CELLS: [[[Cell; 9]; 9]; 3] = {
 
     const fn house_cell(shape: Shape, house: Coord, coord: Coord) -> Cell {
         match shape {
-            Shape::Row => Cell::new(9 * house.u32() + coord.u32()),
-            Shape::Column => Cell::new(house.u32() + 9 * coord.u32()),
-            Shape::Block => Cell::new((house.u32() / 3) * 27 + (house.u32() % 3) * 3 + (coord.u32() / 3) * 9 + (coord.u32() % 3)),
+            Shape::Row => Cell::new(9 * house.u8() + coord.u8()),
+            Shape::Column => Cell::new(house.u8() + 9 * coord.u8()),
+            Shape::Block => Cell::new((house.u8() / 3) * 27 + (house.u8() % 3) * 3 + (coord.u8() / 3) * 9 + (coord.u8() % 3)),
         }
     }
